--- packages/postgres-database/src/simcore_postgres_database/migration/versions/43b26b6e762b_ondelete_in_user_to_project_and_.py
@@ -1,36 +1,0 @@
-"""OnDelete in user_to_project and UniqueConstraint in comp_tasks
-
-Revision ID: 43b26b6e762b
-Revises: 46c3cc232c0b
-Create Date: 2020-02-27 15:49:36.518648+00:00
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '43b26b6e762b'
-down_revision = '46c3cc232c0b'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_unique_constraint('project_node_uniqueness', 'comp_tasks', ['project_id', 'node_id'])
-    op.drop_constraint('user_to_projects_user_id_fkey', 'user_to_projects', type_='foreignkey')
-    op.drop_constraint('user_to_projects_project_id_fkey', 'user_to_projects', type_='foreignkey')
-    op.create_foreign_key(None, 'user_to_projects', 'users', ['user_id'], ['id'], ondelete='CASCADE')
-    op.create_foreign_key(None, 'user_to_projects', 'projects', ['project_id'], ['id'], ondelete='CASCADE')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'user_to_projects', type_='foreignkey')
-    op.drop_constraint(None, 'user_to_projects', type_='foreignkey')
-    op.create_foreign_key('user_to_projects_project_id_fkey', 'user_to_projects', 'projects', ['project_id'], ['id'])
-    op.create_foreign_key('user_to_projects_user_id_fkey', 'user_to_projects', 'users', ['user_id'], ['id'])
-    op.drop_constraint('project_node_uniqueness', 'comp_tasks', type_='unique')
-    # ### end Alembic commands ###
